Correctly handling NAL units in the h264 stream.
This commit adds two functionalities; first of all, the SPS
header is parsed and the picture/sequence parameter sets are
extracted to be then stored in the AVC1 atom.
Second, all the markers before the NAL units are patched and
replaced with the size of the NAL payload, as requested by
ISO 14996-15.